+++ R/model_parameters_default.R
@@ -91,6 +91,30 @@ model_parameters.default <- function(model, ci = .95, bootstrap = FALSE, iterati
 # other special cases ------------------------------------------------
 
 
+#' @export
+model_parameters.averaging <- function(model, ci = .95, component = c("conditional", "full"), exponentiate = FALSE, p_adjust = NULL, ...) {
+  component <- match.arg(component)
+  out <- .model_parameters_generic(
+    model = model,
+    ci = ci,
+    bootstrap = FALSE,
+    iterations = 10,
+    merge_by = "Parameter",
+    standardize = NULL,
+    exponentiate = exponentiate,
+    component = component,
+    effects = "fixed",
+    robust = FALSE,
+    p_adjust = p_adjust,
+    ...
+  )
+
+  attr(out, "object_name") <- deparse(substitute(model), width.cutoff = 500)
+  out
+}
+
+
+
 #' @rdname model_parameters.default
 #' @export
 model_parameters.betareg <- function(model, ci = .95, bootstrap = FALSE, iterations = 1000, component = c("conditional", "precision", "all"), standardize = NULL, exponentiate = FALSE, p_adjust = NULL, ...) {